requirements.txt
altgraph==0.17
appdirs==1.4.3
asgiref==3.3.1
auto-py-to-exe==2.7.11
boto3==1.16.31
botocore==1.19.31
bottle==0.12.19
bottle-websocket==0.2.9
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.6.0
django-json-widget==1.0.1
django-jsonfield==1.4.1
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.2
Eel==0.12.4
et-xmlfile==1.0.1
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==1.0a1
html5lib==1.0.1
idna==2.8
importlib-metadata==3.1.1
install==1.3.4
ipaddr==2.2.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==2.0.2
numpy==1.19.4
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.3
pandas==1.1.5
pep517==0.8.2
phonenumbers==8.12.14
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whichcraft==0.6.1
xlrd==1.2.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
